requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.28.0
pillow>=9.0.0
qrcode>=7.3.1 